my commit file

+++ str_function_2.c
@@ -3,7 +3,7 @@
 char *_strchr(char *s, char c);
 int _strspn(char *s, char *accept);
 int _strcmp(char *s1, char *s2);
-int _strncmp(const char *s1, const char *s2, size_t n);
+int _strncmp(const char *s1, const char *s2, int n);
 
 /**
  * char *_strchr - function that locates a character in a string.
